Binary operator precedence, string checks
Syntactic analyzer modified so that it builds the expression AST considering operator precedence (precedence can be viewed in the resources/operatorprecedence.csv file). String comparisons are validated while building AST (equals and differs are valid, both nodes on each side must be string nodes), raises exception if not valid

Instruction arguments that should not be strings are checked, and, if they are strings, an exception is raised

modified SyntacticAnalizer class so that it can be inyected with firsts map, nexts map, and operator precedence map.