Remaining instructions, execution by mode and steps
Execution also considers breakpoints, terminating execution when encountering a breakpoint with address on pc (that is not the first instruction executed)

annotations have been removed from C memory table and added as hovers. Hover cursor over instruction label to see annotation (if it exists). Instructions with annotations are now written in bold