Execution of instructions part 1
Added Input Dialog TopLevel with validation of input. Cannot close the window without entering value

implemented the part of the main view that displays errors. Both in the Input Dialog and the Main View, to display the errors, a third party component is used: CTkMessagebox. Installed using pip install CTkMessagebox.

Worked with the control panel to make it as 'responsive' as possible. Also:

- Added a spinner to enter n-step number. Entry is validated with each keystroke. It only admits digits, and the spinner only admits to lower the entry down to 1

- Buttons Run, Step, and N-step have their functionalities implemented. Only Step has been tested in this part

Libre, Actual, and PO are updated visually. They are added to the modified cells that correspond to each memory in the virtual machine so the view can update the lines with the general method update_memory.

Not all instructions have been tested.