Output panel and data and heap memory modification
Output panel works as intended: prints SetOut argument

Data and Heap memory View: value is now shown as ### when logical value is None.

Access to data or heap memory (Expression AST) now raises an exception when the value of the data or heap cell has not been initialized. Error dialog explains error and shows address in code memory and line in source file where the instruction that triggered the error is.

Flow of execution is logically followed and visually shown up until the instruction that triggered the exception.