Reset, Upload, NumeralSpinner modifications
Reset now reloads the file if there has been a modification to it. Otherwise, it simply resets view, label panel, and memories.

Data memory and heap memory are now reset logically only if they have been modified

Upload source now resets view if there has been a file uploaded before.

If a load results in a reset/there is a reload (reset is mandatory), the reset will not reset the processor or the label dictionary, since those are set/reset in the load program method

Numeral spinner modified to get its entry value only when clicking the n-step button. Callback is not used.