CodeMemoryView annotation tooltips
List of tooltips so that they are all eventually destroyed and no tooltip is left drawn on the window.

Debounce (of 200ms) made for _on_tree_motion (which schedules the construction of an annotation TopLevel), in new file time.py in view.utils. Deletion of tooltips when leaving the frame is delayed 200ms.

Moved one method to ColorManager (get_non_transparent_color)

Changed add_breakpoint method in BreakpointCanvas so that, if no Breakpoint is passed, it creates a new one with the rest of the arguments (which then are mandatory)