Undo, Save and Save As, Fixes
User can now Undo when there has been an error (test)

Save and Save as are logically implemented, view is partially implemented (shortcuts are set too).

Refactored what was the file managing part of the presenter (time modification checks, other variables that had part in the loading and reloading) in a new class named PresenterFileManager, implemented Save and Save As logic in that class.

Fix: numeral spinner when deleting text and pressing + or - now works (starts at 1)